+++ com.unity.ml-agents/Runtime/Actuators/ActionSegment.cs
@@ -107,6 +107,15 @@ namespace Unity.MLAgents.Actuators
             System.Array.Clear(Array, Offset, Length);
         }
 
+        /// <summary>
+        /// Check if the segment is empty.
+        /// </summary>
+        /// <returns>Whether or not the segment is empty.</returns>
+        public bool IsEmpty()
+        {
+            return Array == null || Array.Length == 0;
+        }
+
         /// <inheritdoc/>
         IEnumerator<T> IEnumerable<T>.GetEnumerator()
         {