+++ ml-agents/mlagents/trainers/env_manager.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+
 from typing import List, Dict, NamedTuple, Iterable, Tuple
 from mlagents_envs.base_env import (
     DecisionSteps,