+++ com.unity.ml-agents/CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to
 ### Major Changes
 #### com.unity.ml-agents (C#)
 #### ml-agents / ml-agents-envs / gym-unity (Python)
-
+- PyTorch trainers now support training agents with both continuous and discrete action spaces. (#4702)
 ### Minor Changes
 #### com.unity.ml-agents / com.unity.ml-agents.extensions (C#)
+- Agents with both continuous and discrete actions are now supported. You can specify
+both continuous and discrete action sizes in Behavior Parameters. (#4702, #4718)
 - In order to improve the developer experience for Unity ML-Agents Toolkit, we have added in-editor analytics.
 Please refer to "Information that is passively collected by Unity" in the
 [Unity Privacy Policy](https://unity3d.com/legal/privacy-policy). (#4677)